+++ xOld/Controller/InboxController.php
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * InterCom Module for Zikula
+ *
+ * @copyright  InterCom Team
+ * @license    GNU/GPL - http://www.gnu.org/copyleft/gpl.html
+ * @package    InterCom
+ * @subpackage User
+ *
+ * Please see the CREDITS.txt file distributed with this source code for further
+ * information regarding copyright.
+ */
+
+namespace Zikula\IntercomModule\Controller;
+
+use Zikula\Core\Controller\AbstractController;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route; // used in annotations - do not remove
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method; // used in annotations - do not remove
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Zikula\IntercomModule\Util\Messages;
+
+/**
+ * @Route("/inbox")
+ *
+ */
+class InboxController extends AbstractController {
+
+ 
+}
